jinson: Hackishly insert a new unknown id in attachment paths
Recent backups from JIRA added another unknown path level in the
attachments tree. Thanks to the fact that JIRA can't even figure the
basic principles of XML (duplicated ids, yay!) and that the export
format is a glorified PostgreSQL dump without foreign key relations,
I haven't figured yet what it is.

All my attachments are in the same group, so this bandaid works for me.